excluding db and gitea access from users

+++ cmd/delete_user.go
@@ -49,17 +49,19 @@ var deleteUserCmd = &cobra.Command{
 			eh.PrintError("finding by hashed email", dberr)
 		}
 
-		gitClient, err := giteautil.NewGitClient()
-		eh.PrintError("instantiating git client", err)
+		if false {
+			gitClient, err := giteautil.NewGitClient()
+			eh.PrintError("instantiating git client", err)
 
-		err = gitClient.DeleteUserRepo(user.Username)
-		eh.PrintError("deleting user repos from gitea", err)
+			err = gitClient.DeleteUserRepo(user.Username)
+			eh.PrintError("deleting user repos from gitea", err)
 
-		err = gitClient.DeleteUserPublicKey(user, user.PublicKeyID)
-		eh.PrintError("deleting user public key from gitea", err)
+			err = gitClient.DeleteUserPublicKey(user, user.PublicKeyID)
+			eh.PrintError("deleting user public key from gitea", err)
 
-		err = gitClient.RemoveUser(user.Username)
-		eh.PrintError("removing user from gitea", err)
+			err = gitClient.RemoveUser(user.Username)
+			eh.PrintError("removing user from gitea", err)
+		}
 
 		output, err := executil.Execute("kubectl delete ingress vscode-"+user.Username, debug)
 		eh.PrintErrorWithOutput("deleting ingress", err, output)
@@ -73,30 +75,32 @@ var deleteUserCmd = &cobra.Command{
 		output, err = executil.Execute("kubectl delete secret vscode-secrets-"+user.Username, debug)
 		eh.PrintErrorWithOutput("deleting secret", err, output)
 
-		registryClient, err := regutil.NewRegistryClient(debug)
-		eh.PrintError("getting registry client", err)
+		if false {
+			registryClient, err := regutil.NewRegistryClient(debug)
+			eh.PrintError("getting registry client", err)
 
-		repository := "curiosinauts/vscode-" + user.Username
-		tags, err := registryClient.ListTags(repository, debug)
-		eh.PrintError("listing tags", err)
+			repository := "curiosinauts/vscode-" + user.Username
+			tags, err := registryClient.ListTags(repository, debug)
+			eh.PrintError("listing tags", err)
 
-		for _, tag := range tags {
-			err = registryClient.DeleteImage(repository, tag, debug)
-			eh.PrintError("deleting image", err)
+			for _, tag := range tags {
+				err = registryClient.DeleteImage(repository, tag, debug)
+				eh.PrintError("deleting image", err)
+			}
+
+			dockerRegistryHost := viper.Get("docker_registry_host").(string)
+			output, err = sshutil.RemoteSSHExec(dockerRegistryHost, "22", "debian",
+				"sudo rm -rf /var/lib/registry/docker/registry/v2/repositories/curiosinauts/vscode-"+user.Username)
+			eh.PrintErrorWithOutput("deleting docker repo folder", err, output)
+
+			postgresUsername := strings.Replace(user.Username, "-", "", -1)
+			psql := postgresutil.NewPSQLClientForSharedDB()
+			_, err = psql.DropUserSchema(postgresUsername, debug)
+			eh.PrintError("dropping database user schema", err)
+
+			_, err = psql.DropUser(postgresUsername, debug)
+			eh.PrintError("dropping database user", err)
 		}
-
-		dockerRegistryHost := viper.Get("docker_registry_host").(string)
-		output, err = sshutil.RemoteSSHExec(dockerRegistryHost, "22", "debian",
-			"sudo rm -rf /var/lib/registry/docker/registry/v2/repositories/curiosinauts/vscode-"+user.Username)
-		eh.PrintErrorWithOutput("deleting docker repo folder", err, output)
-
-		postgresUsername := strings.Replace(user.Username, "-", "", -1)
-		psql := postgresutil.NewPSQLClientForSharedDB()
-		_, err = psql.DropUserSchema(postgresUsername, debug)
-		eh.PrintError("dropping database user schema", err)
-
-		_, err = psql.DropUser(postgresUsername, debug)
-		eh.PrintError("dropping database user", err)
 
 		dberr = dbs.Del(&user)
 		eh.PrintError("delete user", dberr)
